make codec tlv config fields 4bytes wide

+++ src/include/sof/audio/post_process/lib_api.h
@@ -84,8 +84,8 @@ struct post_process_setup_config {
 };
 
 struct pp_param {
-	uint16_t id;
-	uint16_t size;
+	uint32_t id;
+	uint32_t size;
 	uint32_t data[];
 };
